refactor(orc8r): Segregation of StreamerProvider servicer
